Split version logic between MVC modules
